Extract compatibility functions
