+++ jedihttp/hmaclib.py
@@ -12,25 +12,12 @@
 #    limitations under the License.
 
 
-import sys
 import json
 import hmac
 import hashlib
 import tempfile
 from base64 import b64encode, b64decode
-
-
-if sys.version_info[0] == 3:
-  basestring = str
-  unicode = str
-
-
-def encode_string( value ):
-  return value.encode('utf-8') if isinstance(value, unicode) else value
-
-
-def decode_string(value):
-  return value if isinstance(value, basestring) else value.decode('utf-8')
+from jedihttp.compatibility import encode_string, decode_string, compare_digest
 
 
 def TemporaryHmacSecretFile( secret ):
@@ -106,36 +93,3 @@ class JediHTTPHmacHelper( object ):
 
     return compare_digest( self._GetHmacHeader( headers ),
                            self._Hmac( content ) )
-
-
-# hmac.compare_digest were introduced in python 2.7.7
-if sys.version_info >= ( 2, 7, 7 ):
-  from hmac import compare_digest as SecureStringsEqual
-else:
-  # This is the compare_digest function from python 3.4, adapted for 2.6:
-  # http://hg.python.org/cpython/file/460407f35aa9/Lib/hmac.py#l16
-  #
-  # Stolen from https://github.com/Valloric/ycmd
-  def SecureStringsEqual( a, b ):
-    """Returns the equivalent of 'a == b', but avoids content based short
-    circuiting to reduce the vulnerability to timing attacks."""
-    # Consistent timing matters more here than data type flexibility
-    if not ( isinstance( a, str ) and isinstance( b, str ) ):
-      raise TypeError( "inputs must be str instances" )
-
-    # We assume the length of the expected digest is public knowledge,
-    # thus this early return isn't leaking anything an attacker wouldn't
-    # already know
-    if len( a ) != len( b ):
-      return False
-
-    # We assume that integers in the bytes range are all cached,
-    # thus timing shouldn't vary much due to integer object creation
-    result = 0
-    for x, y in zip( a, b ):
-      result |= ord( x ) ^ ord( y )
-    return result == 0
-
-
-def compare_digest( a, b ):
-  return SecureStringsEqual( a, b )
